requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dash-chat==0.3.0
docstring_parser==0.16
faiss-cpu==1.11.0
Flask==3.0.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.19.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.42.0
nest-asyncio==1.6.0
numpy==2.3.0
oauthlib==3.2.2
packaging==24.2
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.1
streamlit-chat==0.1.1
streamlit-chatbox==1.1.13.post1
streamlit-feedback==0.1.4
streamlit-markdown==1.1.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.0.6
zipp==3.23.0
